restructure connect-wallet
