+++ packages/registry/registry/polkadot-ui/blocks/connect-wallet/demo/page.papi.tsx
@@ -1,6 +1,6 @@
 import { PolkadotProvider } from "@/registry/polkadot-ui/lib/polkadot-provider.papi";
 import { ConnectWalletDemo } from "@/registry/polkadot-ui/blocks/connect-wallet/demo/connect-wallet-demo";
-import { WalletSelect } from "@/registry/polkadot-ui/blocks/connect-wallet/components/connect-wallet.papi";
+import { WalletSelect } from "@/registry/polkadot-ui/blocks/connect-wallet/connect-wallet.papi";
 
 export default function ConnectWalletDedotPage() {
   return (
